% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_census_vectors.R
\name{aggregate_census_vectors}
\alias{aggregate_census_vectors}
\title{Aggregate census vectors}
\usage{
aggregate_census_vectors(data)
}
\arguments{
\item{data}{Data for census vectors, from \code{\link{get_census_vectors_and_children}}}
}
\description{
Automatically aggregate census vectors from multiple geographies, based on the type of data. See Details.
}
\details{
The following types of vectors are supported:
\subsection{Units: Number, Aggregation Type: Additive}{

The \code{value} is derived by summing all of the \code{value}s for each vector, across geographies. The result is a count (e.g. population, households). An additional field, \code{value_proportion}, is derived, describing what proportion of the total count (e.g. of population, households) from the highest parent vector falls within each child vector.
}

\subsection{Units: Percentage (0-100), Aggregation Type: Average}{
}

\subsection{Units: Currency, Aggregation Type: Average}{
}
}
\examples{
get_census_vectors_and_children(
  regions = list(CSD = c("3520005", "3521005")),
  level = "CSD",
  vectors = c("v_CA16_401", "v_CA16_418")
) \%>\%
  aggregate_census_vectors()
}
