% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_population_density.R
\name{aggregate_population_density}
\alias{aggregate_population_density}
\title{Aggregate population density}
\usage{
aggregate_population_density(data)
}
\arguments{
\item{data}{Data for census vectors, from \code{\link{get_census_vectors_and_children}}. Must contain a vector with label "Land area in square kilometres" and one vector containing "Population" in the label.}
}
\description{
Average population density across multiple geographies.
}
\examples{
get_census_vectors_and_children(
  regions = list(CSD = c("3520005", "3521005")),
  level = "CSD",
  vectors = c("v_CA16_401", "v_CA16_407")
) \%>\%
  aggregate_population_density()
}
