% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_aggregation_type.R
\name{derive_aggregation_type}
\alias{derive_aggregation_type}
\title{Derive aggregation type}
\usage{
derive_aggregation_type(data)
}
\arguments{
\item{data}{Data containing a field \code{aggregation}, e.g. from \code{\link[cancensus]{list_census_vectors}}}
}
\description{
Derive \code{aggregation_type} field from data containing an \code{aggregation} field. This is a simplified version of the field, e.g. converting \code{aggregation = "Average of v_CA16_4890"} to \code{aggregation_type = "Average"}
}
\examples{
library(cancensus)

list_census_vectors("CA16") \%>\%
  derive_aggregation_type()
}
