% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_census_vector_order.R
\name{derive_census_vector_order}
\alias{derive_census_vector_order}
\title{Derive census vector order}
\usage{
derive_census_vector_order(data, by_value = FALSE)
}
\arguments{
\item{data}{Census data, aggregated via \code{\link{aggregate_census_vectors}}}

\item{by_value}{Whether the ordering is by the value (TRUE) or by the vector numbering (FALSE). Defaults to FALSE.}
}
\description{
Derives the order census vectors should go in, either by using their vector numbering (semantic ordering) or by their values. Returns \code{label} in the data as an ordered factor.
}
\examples{
# TODO
}
