% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_vectors_and_children.R
\name{get_census_vectors_and_children}
\alias{get_census_vectors_and_children}
\title{Get census vectors and children}
\usage{
get_census_vectors_and_children(
  dataset = "CA16",
  regions = "Regions",
  level,
  vectors,
  terminal_only = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{dataset}{The dataset to query vectors from. Defaults to "CA16", the 2016 Canadian Census.}

\item{regions}{A named list of census regions to retrieve. Names must be valid census aggregation levels.}

\item{level}{The census aggregation level to retrieve, defaults to \code{"Regions"}. One of \code{"Regions"}, \code{"PR"}, \code{"CMA"}, \code{"CD"}, \code{"CSD"}, \code{"CT"}, \code{"DA"}, \code{"EA"} (for 1996), or \code{"DB"} (for 2001-2016).}

\item{vectors}{An R vector containing the vector short codes for the Census vectors to download. Vector short codes can be found via the \code{vector} field of \code{\link{census_vectors}}.}

\item{terminal_only}{Whether to include only terminal vectors (ones without children). Defaults to FALSE.}

\item{quiet}{When TRUE, suppress messages and warnings.}
}
\description{
Get data for census vectors and all of their children (and their childrens' children, etc).
}
\examples{
get_census_vectors_and_children(
  regions = list(CSD = c("3520005", "3521005", "3521010")),
  level = "CSD", vectors = c("v_CA16_404", "v_CA16_548")
)
}
