% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassign_parent_vector.R
\name{reassign_parent_vector}
\alias{reassign_parent_vector}
\title{Reassign parent vector}
\usage{
reassign_parent_vector(data, vectors)
}
\arguments{
\item{data}{Census data, from e.g. \link[cancensus]{list_census_vectors} or \link{get_census_vectors_and_children}}

\item{vectors}{A tibble containing \code{vector} and \code{new_parent_vector} for setting a vector's new parent vector in \code{data}. Can optionally include \code{new_label_short} to also replace the \code{label_short} field.}
}
\description{
Update the parent (and highest parent, if relevant) for supplied vectors. Useful for piping into \code{\link{aggregate_census_vectors}}, which often uses the parent vector as a denominator. Optionally, if \code{new_label_short} is provided in \code{vectors}, the \code{label_short} field of \code{data} will also be replaced.
}
\examples{
# TODO
}
