% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_estimated_median_income.R
\name{aggregate_estimated_median_income}
\alias{aggregate_estimated_median_income}
\title{Aggregate estimated median income}
\usage{
aggregate_estimated_median_income(data)
}
\arguments{
\item{data}{Data containing income buckets and total count income is based on}
}
\description{
Estimate median income using income buckets, rounded to the nearest $1,000.
}
