% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_census_vector.R
\name{plot_census_vector}
\alias{plot_census_vector}
\title{Plot a neighbourhood profile variable, for web}
\usage{
plot_census_vector(
  data,
  prop = TRUE,
  width = 20,
  dollar = FALSE,
  vector_description,
 
    aria_label = glue::glue("Bar chart showing the breakdown of {vector_description}. The data is in the table that follows.")
)
}
\arguments{
\item{data}{Census data, aggregated via \code{\link{aggregate_census_vectors}}}

\item{prop}{Whether the vector to be displayed is a proportion. Defaults to TRUE.}

\item{width}{Passed along to str_wrap for wrapping y-axis labels. Defaults to a width of 20.}

\item{dollar}{Whether the variable shown is in dollars. Defaults to FALSE.}

\item{vector_description}{A short description of the vector, used for forming alt text for the plot.}
}
\description{
Plot a neighbourhood profile variable, for web
}
\examples{
# TODO
}
